Refactor token ABI import paths for better organization and consistency
